# packages/contracts/openapi/control-v1.yaml
...
openapi: 3.1.0
info:
  title: OLO ToolGate Control Plane API
  version: 0.10.0-dev
  description: Tenant-scoped group administration with reviewed configuration changes
    and online current authorization. Signed snapshots describe configuration; execution
    requires a separately signed single-use permit.
security:
- adminAccessToken: []
paths:
...
  /api/control/v1/access/simulate:
    post:
      operationId: simulateEnterpriseAccess
      summary: Read-only evaluation of a complete enterprise request
      security:
      - adminAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseEvaluation
      responses:
        '200':
          description: Authoritative current revision and provenance
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseDecision
        '403':
          description: Simulation privilege required
  /api/control/v1/access/invocations:
    get:
      operationId: listEnterpriseInvocations
      responses:
        '200':
          description: Current scoped enterprise authority
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocationPage
        '403':
          description: Current authority denied
        '409':
          description: Exact revision or immutable evidence conflicts
      parameters:
      - name: after
        in: query
        schema:
          type: string
          maxLength: 128
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
    post:
      operationId: submitEnterpriseInvocation
      security:
      - adminAccessToken: []
      responses:
        '200':
          description: Current authoritative state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocation
        '403':
          description: Access denied
        '409':
          description: Revision or invocation binding conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocationRequest
  /api/control/v1/access/invocations/reserve:
    post:
      operationId: reserveEnterprisePermit
      security:
      - adminAccessToken: []
      responses:
        '200':
          description: Current authoritative state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseReservation
        '403':
          description: Access denied
        '409':
          description: Revision or invocation binding conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseReservationRequest
  /api/control/v1/access/invocations/{id}:
    get:
      operationId: getEnterpriseInvocation
      security:
      - adminAccessToken: []
      responses:
        '200':
          description: Current authoritative state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocation
        '403':
          description: Access denied
        '409':
          description: Revision or invocation binding conflict
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: ../schemas/v1/identifiers.schema.json#/$defs/Identifier
  /api/control/v1/access/invocations/{id}/cancel:
    post:
      operationId: cancelEnterpriseInvocation
      security:
      - adminAccessToken: []
      responses:
        '200':
          description: Current authoritative state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocation
        '403':
          description: Access denied
        '409':
          description: Revision or invocation binding conflict
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: ../schemas/v1/identifiers.schema.json#/$defs/Identifier
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
          pattern: ^"[1-9][0-9]{0,15}"$
  /api/control/v1/access/permits/consume:
    post:
      operationId: consumeEnterprisePermit
      security:
      - deviceCertificate: []
      responses:
        '200':
          description: Current authoritative state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocation
        '403':
          description: Access denied
        '409':
          description: Revision or invocation binding conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterprisePermitConsumption
  /api/control/v1/access/invocations/{id}/device-outcome:
    get:
      operationId: getDeviceEnterpriseOutcome
      description: Own-device outcome metadata for journal recovery; never creates an execution permit.
      security:
      - deviceCertificate: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          maxLength: 128
      responses:
        '200':
          description: Authoritative outcome for the authenticated target device
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocation
        '403':
          description: Another target device or inactive certificate
  /api/control/v1/access/effects/report:
    post:
      operationId: reportEnterpriseEffect
      security:
      - deviceCertificate: []
      responses:
        '200':
          description: Current authoritative state
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseInvocation
        '403':
          description: Access denied
        '409':
          description: Revision or invocation binding conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseEffectReport
...
  /api/control/v1/access/human/catalog:
    post:
      operationId: humanToolCatalog
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseHumanTarget
      responses:
        '200':
          description: Current authorized response
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/endpoint.schema.json#/$defs/LocalToolCatalog
        '403':
          description: Current group grant or target trust denied
  /api/control/v1/access/human/invocations:
    post:
      operationId: invokeHumanTool
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseHumanRequest
      responses:
        '200':
          description: Current authorized response
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseHumanOutcome
        '403':
          description: Current group grant or target trust denied
        '202':
          description: Pending independent operation approval; no dispatch or permit
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseHumanOutcome
  /api/control/v1/access/human/invocations/{id}/result:
    get:
      operationId: humanToolResult
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          maxLength: 128
      responses:
        '200':
          description: Current authorized response
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/endpoint.schema.json#/$defs/RemoteToolResponse
        '403':
          description: Current group grant or target trust denied
  /api/control/v1/vault:
    get:
      operationId: listVaultReferences
      responses:
        '200':
          description: Current group-scoped secret references
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseVaultPage
    post:
      operationId: storeVaultSecret
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseVaultWrite
      responses:
        '201':
          description: Encrypted custody and redacted audit committed
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseVaultStored
...
  /api/control/v1/access/invocations/{id}/reconciliation:
    post:
      operationId: proposeOutcomeReconciliation
      responses:
        '200':
          description: Current scoped enterprise authority
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseReconciliation
        '403':
          description: Current authority denied
        '409':
          description: Exact revision or immutable evidence conflicts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseReconciliationRequest
      parameters:
      - &id229
        name: id
        in: path
        required: true
        schema:
          type: string
          maxLength: 128
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          maxLength: 128
  /api/control/v1/access/reconciliations/{id}:
    get:
      operationId: getOutcomeReconciliation
      responses:
        '200':
          description: Current scoped enterprise authority
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseReconciliation
        '403':
          description: Current authority denied
        '409':
          description: Exact revision or immutable evidence conflicts
      parameters:
      - *id229
  /api/control/v1/access/reconciliations/{id}/decision:
    post:
      operationId: decideOutcomeReconciliation
      responses:
        '200':
          description: Current scoped enterprise authority
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseReconciliation
        '403':
          description: Current authority denied
        '409':
          description: Exact revision or immutable evidence conflicts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseReconciliationDecision
      parameters:
      - *id229
  /api/control/v1/access/shadow:
    post:
      operationId: shadowEnterpriseAccess
      responses:
        '200':
          description: Current scoped enterprise authority
          content:
            application/json:
              schema:
                $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseShadowResult
        '403':
          description: Current authority denied
        '409':
          description: Exact revision or immutable evidence conflicts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/v1/enterprise.schema.json#/$defs/EnterpriseShadowRequest
...
components:
  securitySchemes:
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 token with verified issuer/audience/expiry and signed tenant_id/groups.
        Not a Gateway runtime credential.
    deviceCertificate:
      type: mutualTLS
      description: Direct verified device TLS identity; forwarded identity headers
        never authenticate.